extraced simple interface of the testframework refactored example tests
