+++ plugins/example/src/test/java/de/unistuttgart/iaas/amyassist/amy/plugin/example/HelloWorldRestTest.java
@@ -23,15 +23,14 @@
 
 package de.unistuttgart.iaas.amyassist.amy.plugin.example;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.MatcherAssert.*;
+import static org.hamcrest.Matchers.*;
 
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.WebTarget;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
@@ -48,7 +47,6 @@ import de.unistuttgart.iaas.amyassist.amy.test.TestFramework;
  * @author Leon Kiefer
  */
 @ExtendWith(FrameworkExtension.class)
-@Disabled
 class HelloWorldRestTest {
 
 	@Reference
