+++ testing-framework/src/main/java/de/unistuttgart/iaas/amyassist/amy/test/LoggerProvider.java
@@ -42,30 +42,19 @@ public class LoggerProvider implements ServiceFunction<Logger> {
 
 	private static final String CONTEXT_IDENTIFIER = "class";
 
-	/**
-	 * @see de.unistuttgart.iaas.amyassist.amy.core.di.provider.ServiceProvider#getService(java.util.Map,
-	 *      java.util.Map)
-	 */
 	@Override
 	public Logger getService(Map<Class<?>, ServiceFactory<?>> resolvedDependencies, Map<String, ?> context) {
 		Class<?> cls = (Class<?>) context.get(CONTEXT_IDENTIFIER);
 		return LoggerFactory.getLogger(cls);
 	}
 
-	/**
-	 * @see de.unistuttgart.iaas.amyassist.amy.core.di.provider.ServiceProvider#getDependencies()
-	 */
 	@Override
 	public Collection<Class<?>> getDependencies() {
 		return Collections.emptyList();
 	}
 
-	/**
-	 * @see de.unistuttgart.iaas.amyassist.amy.core.di.provider.ServiceProvider#getRequiredContextIdentifiers()
-	 */
 	@Override
 	public Collection<String> getRequiredContextIdentifiers() {
 		return Collections.singletonList(CONTEXT_IDENTIFIER);
 	}
-
 }
